pyproject.toml
[project]
name = "adventure-agent"
version = "0.1.0"
description = "Adventure agent built with LangGraph"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=1.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langsmith>=0.2.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agent"]
[tool.setuptools.package-dir]
"agent" = "src/agent"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
    # Print statements are used for error logging in exception handlers
    "T201",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
# graph.py has conditional imports for checkpointer configuration
"src/agent/graph.py" = ["E402"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.4.7",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (requires dev server to be running)",
    "langsmith: marks tests that require LangSmith tracing to be enabled",
]
